Blood-report parser python/backend/requirement.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
Pillow==10.0.1
pytesseract==0.3.10
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.1.78
spacy==3.7.2
transformers==4.35.0
torch==2.1.0